cardsHeader: - fix update and Delete modals
- fix styles
