+++ src/widgets/icon-dropdown/iconDropdown.module.scss
@@ -8,3 +8,8 @@
     outline: none;
   }
 }
+
+.dropDownItem {
+  padding: 0;
+  color: var(--color-light-100);
+}
